kalkulasi keputusan

+++ app/Http/Controllers/KriteriaController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Kriteria;
+use App\Models\TPA;
 use Illuminate\Http\Request;
 
 class KriteriaController extends Controller
@@ -14,8 +15,9 @@ class KriteriaController extends Controller
 
     public function index()
     {
-        $kriterias = Kriteria::orderBy('id', 'asc')->get(['id', 'label', 'sifat', 'bobot']);
-        return view('kriteria.index', compact('kriterias'));
+        $kriterias = Kriteria::orderBy('id', 'asc')->get(['id', 'label', 'sifat', 'bobot', 'satuan_ukur', 'is_deletable']);
+        $totalBobot = $kriterias->sum('bobot');
+        return view('kriteria.index', compact('kriterias', 'totalBobot'));
     }
 
     public function create()
@@ -29,17 +31,28 @@ class KriteriaController extends Controller
             'label' => 'required|string|max:255',
             'sifat' => 'required|string|max:255',
             'bobot' => 'required|numeric|min:0|max:1',
+            'satuan_ukur' => 'string|max:255',
         ]);
         try {
-            Kriteria::create([
+            $kriterium = Kriteria::create([
                 'nama' => strtolower(str_replace(' ', '_', $request->label)),
                 'label' => $request->label,
                 'sifat' => $request->sifat,
                 'bobot' => $request->bobot,
+                'satuan_ukur' => $request->satuan_ukur,
+                'is_deletable' => true
             ]);
-            return redirect()->route('kriteria.index')->with('success', 'Kriteria berhasil ditambahkan.');
         } catch (\Exception $e) {
             return redirect()->back()->withInput()->with('error', 'Gagal menambahkan Kriteria: ' . $e->getMessage());
+        } finally {
+            try {
+                foreach (TPA::all() as $tpa) {
+                    $tpa->kriterias()->attach($kriterium->id, ['nilai' => 0]);
+                }
+            } catch (\Exception $e) {
+                return redirect()->back()->withInput()->with('error', 'Gagal menambahkan Kriteria: ' . $e->getMessage());
+            }
+            return redirect()->route('kriteria.index')->with('success', 'Kriteria berhasil ditambahkan.');
         }
     }
 
@@ -59,16 +72,26 @@ class KriteriaController extends Controller
             'label' => 'required|string|max:255',
             'sifat' => 'required|string|max:255',
             'bobot' => 'required|numeric|min:0|max:1',
+            'satuan_ukur' => 'string|max:255',
         ]);
         try {
-            $kriterium->update([
-                'nama' => strtolower(
-                    str_replace(' ', '_', $request->label)
-                ),
-                'label' => $request->label,
-                'sifat' => $request->sifat,
-                'bobot' => $request->bobot,
-            ]);
+            if ($kriterium->is_deletable == false) {
+                $kriterium->update([
+                    'bobot' => $request->bobot,
+                    'satuan_ukur' => $request->satuan_ukur,
+                ]);
+            } else {
+                $kriterium->update([
+                    'nama' => strtolower(
+                        str_replace(' ', '_', $request->label)
+                    ),
+                    'label' => $request->label,
+                    'sifat' => $request->sifat,
+                    'bobot' => $request->bobot,
+                    'satuan_ukur' => $request->satuan_ukur,
+                ]);
+            }
+
             return redirect()->route('kriteria.index')->with('success', 'Kriteria berhasil diperbarui.');
         } catch (\Exception $e) {
             return redirect()->back()->withInput()->with('error', 'Gagal memperbarui Kriteria: ' . $e->getMessage());
@@ -78,6 +101,9 @@ class KriteriaController extends Controller
     public function destroy(Kriteria $kriterium)
     {
         try {
+            if ($kriterium->is_deletable == false) {
+                return redirect()->back()->withInput()->with('error', 'Kriteria tidak dapat dihapus.');
+            }
             $kriterium->delete();
             return redirect()->route('kriteria.index')->with('success', 'Kriteria berhasil dihapus');
         } catch (\Exception $e) {
